requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
boto3==1.36.17
botocore==1.36.17
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
channels==3.0.5
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==3.0.2
Django==5.1.4
dnspython==2.7.0
executing==2.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==21.2.0
httplib2==0.22.0
hyperlink==21.0.0
icecream==2.1.3
idna==3.10
incremental==24.7.2
jmespath==1.0.1
packaging==24.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
s3transfer==0.11.2
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uuid==1.30
whitenoise==6.8.2
zope.interface==7.2
django-cors-headers==4.7.0